refactor ViewPost directory
